+++ CuentasClaras/Api/Stats/TopProductGraph.cs
@@ -3,12 +3,12 @@
     public class TopGraph
     {
         public string Description { get; set; }
-        public double Q1 { get; set; }
-        public double Q3 { get; set; }
-        public double Median { get; set; }
-        public double Mean { get; set; }
-        public double Max { get; set; }
-        public double Min { get; set; }
+        public decimal Q1 { get; set; }
+        public decimal Q3 { get; set; }
+        public decimal Median { get; set; }
+        public decimal Mean { get; set; }
+        public decimal Max { get; set; }
+        public decimal Min { get; set; }
         public string DescriptionLong { get; internal set; }
     }
 }
